// examples/retail-store/migrations/snapshots/41bd5540ea833c009b5eacf6db7d489e88bdd5062554398ba8aa2657fa19d27e/contract.d.ts
export type CodecTypes = MongoCodecTypes;
// ...
export type FieldOutputTypes = {
  readonly __unbound__: {
    readonly AddToCartEvent: {
      readonly brand: CodecTypes['mongo/string@1']['output'];
      readonly productId: CodecTypes['mongo/string@1']['output'];
    };
    readonly Cart: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly items: ReadonlyArray<CartItemOutput>;
      readonly userId: CodecTypes['mongo/objectId@1']['output'];
    };
    readonly Event: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly sessionId: CodecTypes['mongo/string@1']['output'];
      readonly timestamp: CodecTypes['mongo/date@1']['output'];
      readonly type: CodecTypes['mongo/string@1']['output'];
      readonly userId: CodecTypes['mongo/string@1']['output'];
    };
    readonly Invoice: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly issuedAt: CodecTypes['mongo/date@1']['output'];
      readonly items: ReadonlyArray<InvoiceLineItemOutput>;
      readonly orderId: CodecTypes['mongo/objectId@1']['output'];
      readonly subtotal: CodecTypes['mongo/double@1']['output'];
      readonly tax: CodecTypes['mongo/double@1']['output'];
      readonly total: CodecTypes['mongo/double@1']['output'];
    };
    readonly Location: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly city: CodecTypes['mongo/string@1']['output'];
      readonly country: CodecTypes['mongo/string@1']['output'];
      readonly name: CodecTypes['mongo/string@1']['output'];
      readonly postalCode: CodecTypes['mongo/string@1']['output'];
      readonly streetAndNumber: CodecTypes['mongo/string@1']['output'];
    };
    readonly Order: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly items: ReadonlyArray<OrderLineItemOutput>;
      readonly shippingAddress: CodecTypes['mongo/string@1']['output'];
      readonly statusHistory: ReadonlyArray<StatusEntryOutput>;
      readonly type: 'delivery' | 'pickup';
      readonly userId: CodecTypes['mongo/objectId@1']['output'];
    };
    readonly Product: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly articleType: CodecTypes['mongo/string@1']['output'];
      readonly brand: CodecTypes['mongo/string@1']['output'];
      readonly code: CodecTypes['mongo/string@1']['output'];
      readonly description: CodecTypes['mongo/string@1']['output'];
      readonly embedding: ReadonlyArray<CodecTypes['mongo/double@1']['output']> | null;
      readonly image: ImageOutput;
      readonly name: CodecTypes['mongo/string@1']['output'];
      readonly price: PriceOutput;
      readonly primaryCategory: CodecTypes['mongo/string@1']['output'];
      readonly status: 'active' | 'discontinued' | 'out-of-stock';
      readonly subCategory: CodecTypes['mongo/string@1']['output'];
    };
    readonly SearchEvent: { readonly query: CodecTypes['mongo/string@1']['output'] };
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly address: AddressOutput | null;
      readonly email: CodecTypes['mongo/string@1']['output'];
      readonly name: CodecTypes['mongo/string@1']['output'];
    };
    readonly ViewProductEvent: {
      readonly brand: CodecTypes['mongo/string@1']['output'];
      readonly exitMethod: CodecTypes['mongo/string@1']['output'] | null;
      readonly productId: CodecTypes['mongo/string@1']['output'];
      readonly subCategory: CodecTypes['mongo/string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly __unbound__: {
    readonly AddToCartEvent: {
      readonly brand: CodecTypes['mongo/string@1']['input'];
      readonly productId: CodecTypes['mongo/string@1']['input'];
    };
    readonly Cart: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly items: ReadonlyArray<CartItemInput>;
      readonly userId: CodecTypes['mongo/objectId@1']['input'];
    };
    readonly Event: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly sessionId: CodecTypes['mongo/string@1']['input'];
      readonly timestamp: CodecTypes['mongo/date@1']['input'];
      readonly type: CodecTypes['mongo/string@1']['input'];
      readonly userId: CodecTypes['mongo/string@1']['input'];
    };
    readonly Invoice: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly issuedAt: CodecTypes['mongo/date@1']['input'];
      readonly items: ReadonlyArray<InvoiceLineItemInput>;
      readonly orderId: CodecTypes['mongo/objectId@1']['input'];
      readonly subtotal: CodecTypes['mongo/double@1']['input'];
      readonly tax: CodecTypes['mongo/double@1']['input'];
      readonly total: CodecTypes['mongo/double@1']['input'];
    };
    readonly Location: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly city: CodecTypes['mongo/string@1']['input'];
      readonly country: CodecTypes['mongo/string@1']['input'];
      readonly name: CodecTypes['mongo/string@1']['input'];
      readonly postalCode: CodecTypes['mongo/string@1']['input'];
      readonly streetAndNumber: CodecTypes['mongo/string@1']['input'];
    };
    readonly Order: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly items: ReadonlyArray<OrderLineItemInput>;
      readonly shippingAddress: CodecTypes['mongo/string@1']['input'];
      readonly statusHistory: ReadonlyArray<StatusEntryInput>;
      readonly type: 'delivery' | 'pickup';
      readonly userId: CodecTypes['mongo/objectId@1']['input'];
    };
    readonly Product: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly articleType: CodecTypes['mongo/string@1']['input'];
      readonly brand: CodecTypes['mongo/string@1']['input'];
      readonly code: CodecTypes['mongo/string@1']['input'];
      readonly description: CodecTypes['mongo/string@1']['input'];
      readonly embedding: ReadonlyArray<CodecTypes['mongo/double@1']['input']> | null;
      readonly image: ImageInput;
      readonly name: CodecTypes['mongo/string@1']['input'];
      readonly price: PriceInput;
      readonly primaryCategory: CodecTypes['mongo/string@1']['input'];
      readonly status: 'active' | 'discontinued' | 'out-of-stock';
      readonly subCategory: CodecTypes['mongo/string@1']['input'];
    };
    readonly SearchEvent: { readonly query: CodecTypes['mongo/string@1']['input'] };
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly address: AddressInput | null;
      readonly email: CodecTypes['mongo/string@1']['input'];
      readonly name: CodecTypes['mongo/string@1']['input'];
    };
    readonly ViewProductEvent: {
      readonly brand: CodecTypes['mongo/string@1']['input'];
      readonly exitMethod: CodecTypes['mongo/string@1']['input'] | null;
      readonly productId: CodecTypes['mongo/string@1']['input'];
      readonly subCategory: CodecTypes['mongo/string@1']['input'];
    };
  };
};
// ...
export type TypeMaps = MongoTypeMaps<CodecTypes, FieldOutputTypes, FieldInputTypes>;
// ...
export type Contract = MongoContractWithTypeMaps<ContractBase, TypeMaps>;